Move cntk layers to separate file
